[FIX] sale: ensure correct untaxed amount with early payment discount (mixed)
Steps to produce:
- Install the Sales module.
- Go to Invoicing > Configuration > Payment Terms.
- Open `Immediate Payment` and enable Early Discount.
- Set `Reduced Tax` to `Always (upon invoice)`.
- Go to Invoicing > Configuration > Taxes.
- Create a 21% tax with `Tax included`.
- Create a product with sale price 7.50 and assign the tax.
- Create a Sale Order with this product > Set Immediate Payment as payment term.

Issue:
- The untaxed amount is computed as 6.22 instead of 6.20.

Root cause:
- At [1], in `_add_base_lines_for_early_payment_discount`, the base lines
  generated for early payment discount were missing the
  `special_mode='total_excluded'`.
- Consequently, the tax engine interpreted these amounts as
  tax-included and attempted to recompute the untaxed base,
  resulting in an incorrect untaxed amount.

Solution:
- Add `special_mode='total_excluded'` to the base lines created
  for early payment discount computation.
- This ensures the discount amounts are treated as already
  tax-excluded.

[1]: https://github.com/odoo/odoo/blob/a2b4f618328f3ce3f654fd2c1ee4410365706a7e/addons/sale/models/sale_order.py#L515-L550

opw-6023472

closes odoo/odoo#260934

X-original-commit: 37250c8ad430d621fc75c8bdcdf4c8c43ab35f22
