+++ addons/sale/tests/__init__.py
@@ -13,6 +13,7 @@ from . import test_product_catalog
 from . import test_product_configurator_data
 from . import test_sale_combo_configurator
 from . import test_sale_flow
+from . import test_sale_early_payment_discount
 from . import test_sale_order
 from . import test_sale_order_product_catalog
 from . import test_sale_order_ui
